refactor GeneratorConfig into DetectionGeneratorConfig
